chore: separate route tests from unit test structure
