+++ clinic-system/server/tests/routes/health.test.js
@@ -1,5 +1,5 @@
 const request = require('supertest')
-const app = require('../../../src/app')
+const app = require('../../src/app')
 
 describe('Health check', () => {
   test('GET /api returns 200', async () => {
